Promote screen project to top-level module
